Move store in proper dir
